requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
gunicorn==21.2.0
pydantic==1.10.9
psycopg2-binary==2.9.6
pymongo==4.4.1
httpx==0.24.1
python-multipart==0.0.6
starlette==0.27.0
aiohttp==3.8.5
aiodns==3.0.0
cchardet==2.1.7
asyncio==3.4.3 